Extract lobby func from SpawnManager
