+++ core/src/main/java/xiao/battleroyale/api/game/spawn/IGameLobbyManager.java
@@ -1,6 +1,7 @@
 package xiao.battleroyale.api.game.spawn;
 
+import xiao.battleroyale.api.game.IGameSubManager;
 import xiao.battleroyale.api.utilitity.ILobbyFuncApi;
 
-public interface IGameLobbyManager extends IGameLobbyReadApi, ILobbyFuncApi {
+public interface IGameLobbyManager extends IGameSubManager, IGameLobbyReadApi, ILobbyFuncApi {
 }
